test(indexer): keep preservation fixtures schema-current
